Various issues
Fixed problem with not setting the last outlet's pin to output
Added then commented out ability to reset the controller and some
verbose serial info due to lack of memory.

The help menu will not show all of the options due to more then likely
using too much memory. Working on this…